pyproject.toml
[project]
name = "husky-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "convex>=0.7.0",
    "httpx>=0.28.1",
    "ipython>=9.4.0",
    "lxml>=6.0.0",
    "numpy>=2.3.1",
    "psycopg>=3.2.9",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
